Initial commit, with import functionality
The import functionality involves calling a S3 generic on a
symbol specification ('symbol_spec') object or list. The 'symbol_spec'
class is virtual and is meant to be extended by concrete class for each
specific data source.

A 'symbol_spec' list can contain objects from multiple sources, and
the appropriate method for each source will be dispatched based on the
object's class that inherits from 'symbol_spec'. Notice that there are
'symbol_spec' methods for each generic, for this purpose.

We will start implementation by building on directly on top of
quantmod. We will eventually remove quantmod entirely, but this allows
us to get started quickly and focus on iterating on the new
functionality.

Also, the import_ohlc() functions return a 'multiple_series' class
object, which is a list of xts objects. This is not 'ideal', but the
idea is that we should always return() the data and avoid side-effects.
But that presents the problem of how to return data for multiple series
in one function call. Whatever we return, the returned class should be
the same whether we return one or many series.

One last thing that makes me uncomfortable: the c.symbol_spec() method
returns a different class depending on how many objects are passed to
it. I don't like this. It should always return the same type of object.
But this was the easiest way I could create a list of specs to different
sources.