List of symbol_specs, each with all attributes
This puts all symbol_spec objects in a list, even for a set of
symbol_spec that share the same source. Each symbol_spec is in one
list element and contains all the relevant source attributes. All
sym_* functions would return symbol_spec_list objects.

This makes it easy to combine symbol_spec objects from different
sources. But it makes it difficult to dispatch to the underlying
methods in a straightforward way. Dispatch would happen one symbol at
a time, when it would be more efficient to batch all the requests to
each source into one method call per source.