Put source attributes on ticker vector
The symbol_spec is a list of tickers by source. The vector of tickers
contains the connection attributes / parameters and the connection
class.

The downside is that the order of the tickers may change when you
combine vectors, because the tickers are added to each source vector.

Update the print() and c() methods to account for the above change.