Add source attributes to symbol spec vectors
Each symbol specification object now has an attribute that defines
all the necessary connection details for the source.

Update the c() method to combine all the source attributes for each
vector. This allows us to separate the symbols by source, then call
the source import method with the necessary connection details.

The .combine_src_attr() function checks for source connection details
that are different even though they're for the same source. Need to
add a unit test for this.