Add curl_options to function parameters
It will be common for people to need to specify curl options for
internet connectivity. Expose them as function parameters so they're
easier to handle (i.e. you don't have to check for them in '...').

Also put 'api_key' parameter after '...' in sym_tiingo(). That makes
all the sym_* functions consistent, because everything after '...'
has to be fully named.