Update import_ohlc() for new symbol_spec structure
Split symbol_spec vector by source, then loop over the symbols for
each source. This way, we can pull all the data for a source in one
request. This may also make it possible to import from multiple
sources in parallel. Extract the 'src_attr' for each source and attach
it to the object that contains the symbols passed to the method.

I tried to use NextMethod() and manipulate the environment to get this
to work. But I couldn't change any of the arguments passed to the next
method. So each method was called with the object passed to the
generic. That meant that the methods received data they didn't expect.

Use getS3method() to find the method for each source. This needs to be
tested to see if it works if the method is in another package.

Add a c() method for 'multiple_ohlc' objects so we can combine the
results from the call to each source method. Otherwise c() strips all
the attributes.