Initial commit of 'price frame' prototype
This function is meant to provide the same functionality as
    do.call("merge", lapply(xts_object, Cl))

A `multiple_ohlc` object is a list of xts objects, by symbol.
make_xts_frame() will extract the same column from each symbol's xts
object and merge them into a single xts object.

I don't love the 'frame' portion of the name, since it's an xts
object, not a data.frame... but maybe that means we need to actually
create the xts data.frames we've talked about for years.

See #5.