Output results the same as input symbols
We currently load getSymbols() results into an environment and then
convert it to a list. The order of the symbols in the list isn't
guaranteed.

Order the symbols in the result to match the order in the symbol spec
input. Also handle the case when the symbol isn't a valid R name
(e.g. "^GSPC" -> "GSPC").

Fixes #2.