Fix regex to remove symbol from column names
The original regex didn't account for the case where the symbol has
a '.' (e.g. BRK.A). The new regex uses Perl look-ahead assertions to
find the last '.' in the column name, so we can remove everything up
to that last '.'.

Thanks to Ethan B. Smith for catching this issue.

See #4.