import_ohlc() returns xts; add import_multi_ohlc()
I was annoyed that I had to subset the output of import_ohlc() when
importing a single symbol, so I wanted a function that returns an xts
object when only one symbol is provided.

Rename import_ohlc() to import_multi_ohlc() because it will always
return a multiple_ohlc object. Create a new import_ohlc() that only
accepts a single symbol and returns an xts object.

See #9.