Add 'dates' argument to generic import functions
Allow users to specify the start and end times in one ISO-8601 style
string. This is very convenient during interactive use.

Add a function to parse the 'dates' argument, and put it in a new
'api' environment that will be exported for developers to use when
writing new source methods in other packages.

See #10.